Move get_job_schema_by_cloud to controller module.
To better organize helper methods which will be used by CLI and
can be interfaced from code as an SDK.
